+++ ninja-servlet/src/main/java/ninja/servlet/ContextImpl.java
@@ -322,6 +322,11 @@ public class ContextImpl implements Context.Impl {
     @Override
     public List<Cookie> getCookies() {
         javax.servlet.http.Cookie[] servletCookies = httpServletRequest.getCookies();
+        
+        if (servletCookies == null) {
+            return Collections.EMPTY_LIST;
+        }
+        
         List<Cookie> ninjaCookies = new ArrayList<>(servletCookies.length);
 
         for (javax.servlet.http.Cookie cookie : servletCookies) {
@@ -633,6 +638,22 @@ public class ContextImpl implements Context.Impl {
     @Override
     public void setAttribute(String name, Object value) {
         httpServletRequest.setAttribute(name, value);
+    }
+
+    @Override
+    public Map<String, Object> getAttributes() {
+        // build map of attributes
+        Map<String,Object> attributes = new HashMap<String,Object>();
+        
+        Enumeration<String> en = httpServletRequest.getAttributeNames();
+        
+        while (en.hasMoreElements()) {
+            String name = en.nextElement();
+            Object value = httpServletRequest.getAttribute(name);
+            attributes.put(name, value);
+        }
+        
+        return attributes;
     }
 
     /**