+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -3,6 +3,9 @@ Version 5.1.7
 
  * 2015-10-01 Fixed garbled snippets in diagnostics (mallowlabs)
  * 2015-09-25 Upgrade FreeMarker from 2.3.22 to 2.3.23 (khmarbaise)
+ * 2015-09-28 Improved diagnostic mode w/ info about the context, request, and response (jjlauer)
+ * 2015-09-28 Fixed NPE issue with cookies not being set in underlying servlet request (jjlauer)
+ * 2015-09-28 NinjaDefault now logs exceptions in the debug level (jjlauer)
 
 Version 5.1.6
 =============